Move /static build into /server builds
This joins the static (prerender) builds with the server builds.
