+++ packages/react-dom/npm/static.browser.js
@@ -1,7 +1,11 @@
 'use strict';
 
+var s;
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-dom-static.browser.production.min.js');
+  s = require('./cjs/react-dom-server.browser.production.min.js');
 } else {
-  module.exports = require('./cjs/react-dom-static.browser.development.js');
+  s = require('./cjs/react-dom-server.browser.development.js');
 }
+
+exports.version = s.version;
+exports.prerender = s.prerender;
