+++ packages/react-dom/npm/static.node.js
@@ -1,7 +1,11 @@
 'use strict';
 
+var s;
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-dom-static.node.production.min.js');
+  s = require('./cjs/react-dom-server.node.production.min.js');
 } else {
-  module.exports = require('./cjs/react-dom-static.node.development.js');
+  s = require('./cjs/react-dom-server.node.development.js');
 }
+
+exports.version = s.version;
+exports.prerenderToNodeStream = s.prerenderToNodeStream;
