+++ packages/react-dom/src/server/react-dom-server.browser.js
@@ -8,3 +8,4 @@
  */
 
 export * from './ReactDOMFizzServerBrowser.js';
+export {prerender} from './ReactDOMFizzStaticBrowser.js';
