+++ packages/react-dom/src/server/react-dom-server.edge.js
@@ -8,3 +8,4 @@
  */
 
 export * from './ReactDOMFizzServerEdge.js';
+export {prerender} from './ReactDOMFizzStaticEdge.js';
