+++ packages/react-dom/src/server/react-dom-server.node.js
@@ -8,3 +8,4 @@
  */
 
 export * from './ReactDOMFizzServerNode.js';
+export {prerenderToNodeStream} from './ReactDOMFizzStaticNode.js';
