+++ packages/react-dom/static.browser.js
@@ -7,4 +7,4 @@
  * @flow
  */
 
-export {prerender, version} from './src/server/ReactDOMFizzStaticBrowser';
+export {prerender, version} from './src/server/react-dom-server.browser';
